Refactor splitting out a privileged server which manages local directories, credentials, permissions and access control.  It receives commands through a token backed messaging over a UNIX domain socket.
The main concept, is each service can have multiple docker mounts.
A root for all service mounts is specified, and each service gets it's
own sub directory.  Each service mount also has it's own directory and
may be versioned.  Service mount versions may be changed as needed.  A
symlink folder, named 'current' will be used to indicate which is the
current version.  Each service gets its own user and group created and
assigned to each folder, while the service root runs under it's own
credentials.  Additionally, all service accounts are added to a
'douglas' group for easy identification.  All user and service group
names are prefixed with 'doug-' and limited to 28 characters.  Folder
names are normalized to uri encoding but allowing minus, period and
underscore

So the directory structure looks something like this:

mount-root
  service-a
    mount-a
      current -> mount-root/service-a/mount-a/v1
      v0
      v1
    mount-b
      current  -> mount-root/service-b/mount-a/v0
      v0
  service-b
    mount-a
      current  -> mount-root/service-b/mount-a/v0
      v0

Mount root is owned by the service root (doug-radilce), while the
service mount directories are owned by their users and groups.
Currently the functionality is limited to UNIX-y systems, this could
potentially work with Windows, but perhaps for a later version.

Server commands:
* ActiveMountVersion
Displays the current version of a service mount that is active.

* CreateCredentials
Creates a user name and group name based on the service name.  This must
be done first, before creating a mount.

* CreateMount
For existing credentials, creates a new mount, initializing it to
version 0.

* CreateNewMountVersion
For an existing mount, creates a new version, and updates the current
version to that.

* ListMountVersions
Lists all the available versions for a current service mount.

* SetMountVersion
For a given service mount, changes the active version.